refactor: split auth client config and providers
