+++ src/auth.ts
@@ -1,27 +1,73 @@
 import NextAuth from "next-auth";
-import authConfig from "./auth.config";
+import authClientConfig from "./auth.client-config";
 import { FirestoreAdapter } from "@auth/firebase-adapter";
 import { admin, adminAuth, adminDb } from "./firebase/firebaseAdmin";
+import Credentials from "next-auth/providers/credentials";
+import { OAuthProviders } from "./auth.providers";
+
+async function getUserById(userId: string) {
+  try {
+    const userDoc = await adminDb.collection("users").doc(userId).get();
+
+    if (!userDoc.exists) {
+      console.log("No user found with the given ID.");
+      return null;
+    }
+
+    // Access the user data
+    const userData = userDoc.data();
+    return userData;
+  } catch (error) {
+    console.error("Error getting user document:", error);
+    throw error;
+  }
+}
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
-  pages: {
-    signIn: "/auth/sign-in",
-  },
+  ...authClientConfig,
+  providers: [
+    ...OAuthProviders,
+    Credentials({
+      name: "Email Link",
+      credentials: {
+        email: { label: "Email", type: "text" },
+        accessToken: { label: "Access Token", type: "text" },
+      },
+      authorize: async (credentials) => {
+        const { email, accessToken } = credentials;
+
+        try {
+          const decodedToken = await admin
+            .auth()
+            .verifyIdToken(accessToken as string);
+
+          if (decodedToken && decodedToken.email === email) {
+            const user = await getUserById(decodedToken.uid);
+
+            if (user) {
+              return {
+                id: decodedToken.uid,
+                name: user.name,
+                email: user.email,
+              };
+            } else {
+              throw new Error("User not found.");
+            }
+          } else {
+            throw new Error("Invalid credentials.");
+          }
+        } catch (error) {
+          console.error("Failed to decode the credit token", error);
+          throw new Error("Invalid credentials.");
+        }
+      },
+    }),
+  ],
   session: {
     strategy: "jwt",
   },
   callbacks: {
-    authorized({ auth, request: { nextUrl } }) {
-      const isLoggedIn = !!auth?.user;
-      const isOnChat = nextUrl.pathname.startsWith("/chat");
-      if (isOnChat) {
-        if (isLoggedIn) return true;
-        return false; // Redirect unauthenticated users to login page
-      } else if (isLoggedIn) {
-        return Response.redirect(new URL("/chat", nextUrl));
-      }
-      return true;
-    },
+    ...authClientConfig.callbacks,
     session: async ({ session, token }) => {
       if (session?.user) {
         if (token.sub) {
@@ -64,5 +110,4 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
     },
   },
   adapter: FirestoreAdapter(adminDb),
-  ...authConfig,
 });
